removed mapping updater
